+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Arbitrary.hs
@@ -652,8 +652,9 @@ instance Era era => Arbitrary (ShelleyPpupPredFailure era) where
   arbitrary = genericArbitraryU
   shrink = genericShrink
 
-instance (Arbitrary (TxCert era), Era era) => Arbitrary (ShelleyPoolPredFailure era) where
+instance Era era => Arbitrary (ShelleyPoolPredFailure era) where
   arbitrary = genericArbitraryU
+  shrink = genericShrink
 
 instance
   ( Era era